Update CreateDataset.py
Add parser for command line datasets configuration. Add error target to use as a parameter. Now the dataset created is perfectly balanced with best, good, bad and worst type of sequences. Good and bad sequence are taken using a default error target of 4.0 relating the entire tour. This error can be changed on every simulation to create different datasets.